Fix logic when handling possible doublicates (i.e. rd.lvm.lv)
The logic should be: "allowed duplicates _OR_ executes (starts with '$') should simply be appended". In the current code, 'rd.lvm.lv' is in allow_multi, but does not start with '$', so a replace will be tried. At first we still hit the append, because it does not exists, after that, that last occurrence of 'rd.lvm.lv' will made it to the string.